refactor(protocol): split app connector contracts
