+++ packages/protocol/src/app-connector/schemas.ts
@@ -1,0 +1,2 @@
+export * from "./definition.js";
+export * from "./installation.js";
